requirements.txt
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
seaborn
plotly
requests
joblib
streamlit
tabulate
google-generativeai